update all permission on general master data

+++ app/Http/Controllers/Area/AreaController.php
@@ -7,6 +7,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\AreaRequest;
 use App\Models\Area;
 use App\Models\Branch;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\View\View;
@@ -24,16 +25,23 @@ use Illuminate\View\View;
         return view('pages.general-master-data.area.index');
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function data(): JsonResponse
     {
+        $this->authorize('view', Area::class);
         $area = Area::with('branch')->withCount('areaHasUser')->paginate(10);
         return response()->json($area);
     }
 
-    public function search(Request $request)
+    /**
+     * @throws AuthorizationException
+     */
+    public function search(Request $request): JsonResponse
     {
+        $this->authorize('view', Area::class);
         $search = $request->input('search');
-
         $area = Area::with('branch')->when(!empty($search), function ($query) use ($search) {
             $query->where('name', 'like', '%' . $search . '%')
                 ->orWhereHas('branch', function ($query) use ($search) {
@@ -47,6 +55,7 @@ use Illuminate\View\View;
 
     public function store(AreaRequest $request): JsonResponse
     {
+        $this->authorize('create', Area::class);
         Area::create($request->validated());
         return response()->json(['message' => 'Data berhasil ditambahkan.']);
     }
@@ -54,29 +63,47 @@ use Illuminate\View\View;
 
     public function edit(Area $area): JsonResponse
     {
+        $this->authorize('update', $area);
         return response()->json($area);
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function update(AreaRequest $request, Area $area): JsonResponse
     {
+        $this->authorize('update', $area);
         $area->update($request->validated());
         return response()->json(['message' => 'Data berhasil diubah.']);
     }
 
 
+    /**
+     * @throws AuthorizationException
+     */
     public function getBranchData(Request $request): JsonResponse
     {
+        $this->authorize('create', Area::class);
+        $this->authorize('update', Area::class);
         $branch = $this->branch->getData($request);
         return response()->json($branch);
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function selectedBranch(Area $area): JsonResponse
     {
+        $this->authorize('update', $area);
         return response()->json($this->branch->getSelectedData($area->branch_id));
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function destroy(Request $request, Area $area): JsonResponse
     {
+        $this->authorize('delete', $area);
         $implodeID = implode(',', $request->get('id'));
         $explodeID = explode(',', $implodeID);
         $area->whereIn('id', $explodeID)->delete();
@@ -87,8 +114,12 @@ use Illuminate\View\View;
     }
 
 
+    /**
+     * @throws AuthorizationException
+     */
     public function detail(Area $area): View
     {
+        $this->authorize('viewDetail', $area);
         return view('pages.general-master-data.area.detail.index', compact('area'));
     }
